requirements.txt
# Base requirements from v1.0
flask==3.0.0
flask-cors==4.0.0
elasticsearch==8.11.1
google-cloud-aiplatform==1.38.1
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0

# v2.0 Enhanced requirements

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# Redis
redis==5.0.1
hiredis==2.2.3

# Message Queue (A2A Protocol)
pika==1.3.2
celery==5.3.4

# MCP Protocol
mcp==1.16.0

# Payment Processing
stripe==7.8.0

# API Integrations (using mock data for demo)
# amadeus==8.1.0  # Uncomment when integrating real APIs

# Security
cryptography==41.0.7
pyjwt==2.8.0
python-jose==3.3.0

# Async support
aiohttp==3.9.1
asyncio==3.4.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx>=0.27.1

# Monitoring
prometheus-client==0.19.0

# Utilities
pydantic>=2.11.0,<3.0.0
python-multipart>=0.0.9
websockets==12.0
